feat: response를 담당하는 클래스 생성
